// src/canister/satsman/service.did.d.ts
export interface Account {
  'reward_from_referral_in_current_block' : number,
  'minted_rune_in_current_block' : number,
  'withdraw_txid' : [] | [string],
  'price_in_current_block' : number,
  'total_paid_sats' : bigint,
  'total_reward_from_referral' : number,
  'last_update_block' : number,
  'total_minted_rune_amount' : number,
  'address' : string,
  'sats_balance' : bigint,
  'tune_in_current_block' : number,
  'total_referral_reward' : number,
  'withdrawn' : boolean,
  'paid_sats_in_current_block' : bigint,
  'total_avg_price' : number,
  'referral_reward_in_current_block' : number,
}
export interface BlockAggregateData {
  'total_paying_sats' : bigint,
  'total_minted_rune' : bigint,
  'total_ongoing_launch_pools' : number,
  'total_paying_users' : number,
}
export interface BlockState {
  'minted_rune_in_current_block' : bigint,
  'price_in_current_block' : number,
  'total_paid_sats' : bigint,
  'auction_raised_amount_in_current_block' : number,
  'total_exchange_fee' : number,
  'total_minted_rune' : bigint,
  'total_auction_raised_amount' : number,
  'user_accounts' : Array<[string, Account]>,
  'total_deposit_btc_balances' : bigint,
  'total_referral_reward' : number,
  'avg_price' : number,
  'block_height' : number,
  'paid_sats_in_current_block' : bigint,
  'exchange_fee_in_current_block' : number,
  'referral_reward_in_current_block' : number,
}
// ...
export type Filter = { 'Ongoing' : null } |
  { 'Completed' : null } |
  { 'Upcoming' : null };
export interface GetPoolInfoArgs { 'pool_address' : string }
export interface HomePageBlockData {
  'block_second_timestamp' : bigint,
  'ongoing_launches_count' : number,
  'total_release_rune_amount' : bigint,
}
export interface HomePageSatsmanData {
  'total_btc_raised' : bigint,
  'total_satsman_count' : number,
}
export interface IncomeDistributionItem {
  'label' : string,
  'address' : string,
  'percentage' : number,
}
// ...
export interface LaunchPlan {
  'start_height' : number,
  'launch_token' : string,
  'social_info' : SocialInfo,
  'rune_name' : string,
  'banner' : [] | [string],
  'description' : [] | [string],
  'is_fair_launch' : boolean,
  'span_blocks' : number,
  'raising_target_sats' : bigint,
  'extract_auction_income_ratio' : number,
  'rune_id' : string,
  'income_distribution' : Array<IncomeDistributionItem>,
}
export interface LaunchpadState {
  'user_balances' : Array<[string, bigint]>,
  'withdrawn_rune' : boolean,
  'txid' : string,
  'utxo' : Utxo,
  'event' : Event,
  'nonce' : bigint,
  'withdrawn_user_set' : Array<string>,
}
export interface NewBlockInfo {
  'block_hash' : string,
  'confirmed_txids' : Array<string>,
  'block_timestamp' : bigint,
  'block_height' : number,
}
export type Outcome = { 'Failed' : null } |
  { 'Listed' : null } |
  { 'Success' : null };
// ...
export interface PoolBusinessStateView {
  'status' : Filter,
  'creator' : string,
  'featured' : boolean,
  'start_height' : number,
  'pool_name' : string,
  'end_height' : number,
  'income_distribution_list' : Array<[string, bigint]>,
  'pubkey' : string,
  'btc_amount_for_lp' : bigint,
  'raising_target' : bigint,
  'key_derivation_path' : Array<Uint8Array | number[]>,
  'exchange_fee_percentage' : number,
  'referral_bonus_percentage' : number,
  'distribute_income_txid' : [] | [string],
  'highest_block_states' : [] | [BlockState],
  'creator_withdraw_rune_txid' : [] | [string],
  'total_claimed_rune' : bigint,
  'total_claimed_sats' : bigint,
  'synced_block_height' : number,
  'user_tunes' : Array<[string, number]>,
  'rune_amount_for_lp' : bigint,
  'rune_amount_for_launch' : bigint,
  'pool_address' : string,
  'outcome' : Outcome,
  'launch_plan' : LaunchPlan,
  'rune_id' : string,
}
// ...
export interface SocialInfo {
  'twitter' : [] | [string],
  'website' : [] | [string],
  'discord' : [] | [string],
  'telegram' : [] | [string],
  'github' : [] | [string],
}
// ...
export declare const idlFactory: IDL.InterfaceFactory;